requirements.txt
asgiref==3.8.1
b64==0.4
beautifulsoup4==4.12.3
bs4==0.0.2
cached_properties==0.7.4
certifi==2024.7.4
charset-normalizer==3.3.2
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
idna==3.7
lxml==5.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2